feat: improve profile loading and localize article card fields
